+++ app/src/main/AndroidManifest.xml
@@ -15,13 +15,13 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.MaterialComponents.NoActionBar">
         <activity
-            android:name=".View.FavouritesActivity"
+            android:name=".view.FavouritesActivity"
             android:exported="false" />
         <activity
-            android:name=".View.SpecificMovie"
+            android:name=".view.SpecificMovie"
             android:exported="false" />
         <activity
-            android:name=".View.SplashScreen"
+            android:name=".view.SplashScreen"
             android:exported="true"
             android:screenOrientation="portrait"
             android:theme="@style/Theme.MaterialComponents.NoActionBar.Fullscreen">
@@ -32,7 +32,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".View.MainActivity"
+            android:name=".view.MainActivity"
             android:exported="true"
             android:screenOrientation="sensor" />
     </application>
