+++ app/src/main/java/com/example/moviebooking/model/PopularMoviesModel.kt
@@ -1,4 +1,4 @@
-package com.example.moviebooking.Model
+package com.example.moviebooking.model
 
 
 import com.google.gson.annotations.SerializedName
