+++ app/src/main/java/com/example/moviebooking/model/Result.kt
@@ -1,23 +1,17 @@
-package com.example.moviebooking.Model
+package com.example.moviebooking.model
 
 
 import com.google.gson.annotations.SerializedName
 
-data class SpecificMovieModel(
+data class Result(
     @SerializedName("adult")
     val adult: Boolean,
     @SerializedName("backdrop_path")
     val backdropPath: String,
-    @SerializedName("belongs_to_collection")
-    val belongsToCollection: Any,
-    @SerializedName("budget")
-    val budget: Int,
-    @SerializedName("homepage")
-    val homepage: String,
+    @SerializedName("genre_ids")
+    val genreIds: List<Int>,
     @SerializedName("id")
     val id: Int,
-    @SerializedName("imdb_id")
-    val imdbId: String,
     @SerializedName("original_language")
     val originalLanguage: String,
     @SerializedName("original_title")
@@ -27,18 +21,9 @@ data class SpecificMovieModel(
     @SerializedName("popularity")
     val popularity: Double,
     @SerializedName("poster_path")
-    val posterPath: Any,
-
+    val posterPath: String,
     @SerializedName("release_date")
     val releaseDate: String,
-    @SerializedName("revenue")
-    val revenue: Int,
-    @SerializedName("runtime")
-    val runtime: Int,
-    @SerializedName("status")
-    val status: String,
-    @SerializedName("tagline")
-    val tagline: String,
     @SerializedName("title")
     val title: String,
     @SerializedName("video")
