+++ app/src/main/java/com/example/moviebooking/network/RetrofitHelper.kt
@@ -1,0 +1,52 @@
+package com.example.moviebooking.network
+
+import com.example.moviebooking.utils.Credentials
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import okhttp3.*
+import retrofit2.Retrofit
+import retrofit2.converter.gson.GsonConverterFactory
+import java.util.concurrent.TimeUnit
+import javax.inject.Singleton
+
+@Module
+@InstallIn(SingletonComponent::class)
+class RetrofitHelper {
+
+    private val requestInterceptor = Interceptor { chain ->
+        val url: HttpUrl = chain.request()
+            .url()
+            .newBuilder()
+            .addQueryParameter("api_key", com.example.moviebooking.utils.Credentials.API_KEY)
+            .build()
+
+        val request: Request = chain.request()
+            .newBuilder()
+            .url(url)
+            .build()
+
+        return@Interceptor chain.proceed(request)
+    }
+
+    private val okHttpClient = OkHttpClient.Builder()
+        .addInterceptor(requestInterceptor)
+        .connectTimeout(60, TimeUnit.SECONDS)
+        .build()
+
+    @Provides
+    @Singleton
+    fun getMovieServiceInstance(retrofit: Retrofit): MovieService =
+        retrofit.create(MovieService::class.java)
+
+    @Provides
+    @Singleton
+    fun getInstance(): Retrofit {
+        return Retrofit.Builder()
+            .client(okHttpClient)
+            .baseUrl(Credentials.BASE_URL)
+            .addConverterFactory(GsonConverterFactory.create())
+            .build()
+    }
+}
