+++ app/src/main/java/com/example/moviebooking/repository/SpecificMovieRepository.kt
@@ -1,10 +1,10 @@
-package com.example.moviebooking.Repository
+package com.example.moviebooking.repository
 
 import android.util.Log
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
-import com.example.moviebooking.Model.SpecificMovieModel
-import com.example.moviebooking.Network.MovieService
+import com.example.moviebooking.model.SpecificMovieModel
+import com.example.moviebooking.network.MovieService
 import javax.inject.Inject
 
 class SpecificMovieRepository @Inject constructor(private val movieService : MovieService) {
@@ -19,6 +19,7 @@ class SpecificMovieRepository @Inject constructor(private val movieService : Mov
     suspend fun getSpecificMovieDetails(movieId : Int)
     {
         val result = movieService.getMovieDetails(movieId)
+        Log.d("kkk","result of repo : "+result.toString())
         if(result?.body() != null)
         {
             Log.d("kkk","specific movie response : "+result.body())
