+++ app/src/main/java/com/example/moviebooking/utils/Credentials.kt
@@ -1,4 +1,4 @@
-package com.example.moviebooking.Utils
+package com.example.moviebooking.utils
 
 class Credentials {
      companion object {
