+++ app/src/main/java/com/example/moviebooking/view/SpecificMovie.kt
@@ -1,28 +1,20 @@
-package com.example.moviebooking.View
+package com.example.moviebooking.view
 
-import android.content.Intent
-import android.graphics.Color
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.util.Log
 import android.widget.ImageView
 import android.widget.RatingBar
 import android.widget.TextView
-import android.widget.Toolbar
 import androidx.lifecycle.ViewModelProvider
-import androidx.room.Room
 import com.bumptech.glide.Glide
 import com.bumptech.glide.load.engine.DiskCacheStrategy
-import com.example.moviebooking.Network.MovieService
-import com.example.moviebooking.Network.RetrofitHelper
 import com.example.moviebooking.R
-import com.example.moviebooking.Repository.SpecificMovieRepository
-import com.example.moviebooking.Utils.Credentials
-import com.example.moviebooking.ViewModel.SpecificMovieViewModel
-import com.example.moviebooking.ViewModelFactory.SpecificMovieViewModelFactory
-import com.bumptech.glide.request.RequestOptions
-import com.example.moviebooking.Database.MovieDatabase
-import com.example.moviebooking.Entity.MovieLike
+import com.example.moviebooking.repository.SpecificMovieRepository
+import com.example.moviebooking.utils.Credentials
+import com.example.moviebooking.viewModel.SpecificMovieViewModel
+import com.example.moviebooking.database.MovieDatabase
+import com.example.moviebooking.entity.MovieLike
 import com.google.android.material.appbar.CollapsingToolbarLayout
 import com.google.android.material.floatingactionbutton.FloatingActionButton
 import dagger.hilt.android.AndroidEntryPoint
