+++ app/src/main/java/com/example/moviebooking/viewModel/PopularMoviesViewModel.kt
@@ -1,10 +1,10 @@
-package com.example.moviebooking.ViewModel
+package com.example.moviebooking.viewModel
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.moviebooking.Model.PopularMoviesModel
-import com.example.moviebooking.Repository.PopularMoviesRepository
+import com.example.moviebooking.model.PopularMoviesModel
+import com.example.moviebooking.repository.PopularMoviesRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
@@ -14,17 +14,12 @@ import javax.inject.Inject
 class PopularMoviesViewModel @Inject constructor (private val popularMoviesRepo : PopularMoviesRepository)  :
     ViewModel() {
 
-
-
     fun getPopularMovies(pageCount : Int)
     {
         viewModelScope.launch(Dispatchers.IO) {
             popularMoviesRepo.getPopularMovies(pageCount)
         }
     }
-
-
-
 
     val popularMoviesLiveData : LiveData<PopularMoviesModel>
     get() = popularMoviesRepo.popularMovieLiveData
