+++ app/src/main/java/com/example/moviebooking/viewModelFactory/SpecificMovieViewModelFactory.kt
@@ -1,10 +1,9 @@
-package com.example.moviebooking.ViewModelFactory
+package com.example.moviebooking.viewModelFactory
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
-import com.example.moviebooking.Repository.SpecificMovieRepository
-import com.example.moviebooking.ViewModel.PopularMoviesViewModel
-import com.example.moviebooking.ViewModel.SpecificMovieViewModel
+import com.example.moviebooking.repository.SpecificMovieRepository
+import com.example.moviebooking.viewModel.SpecificMovieViewModel
 
 class SpecificMovieViewModelFactory(private val specificMovieDetailsRepo : SpecificMovieRepository) : ViewModelProvider.Factory
 {
