+++ app/src/main/res/layout/activity_favourites.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".View.FavouritesActivity">
+    tools:context=".view.FavouritesActivity">
 
 
     <androidx.appcompat.widget.Toolbar
