+++ app/src/main/res/layout/activity_main.xml
@@ -7,7 +7,7 @@
     android:layout_height="match_parent"
     android:id="@+id/linearLayout"
     android:orientation="vertical"
-    tools:context=".View.MainActivity">
+    tools:context=".view.MainActivity">
 
 
     <androidx.appcompat.widget.Toolbar
