+++ app/src/main/res/layout/activity_specific_movie.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".View.SpecificMovie">
+    tools:context=".view.SpecificMovie">
